+++ reprounzip/reprounzip/common.py
@@ -94,11 +94,11 @@ def load_config(filename, canonical, File=File, Package=Package):
 
     ver = LooseVersion(config['version'])
 
-    keys_ = set(config.keys())
+    keys_ = set(config)
     if 'version' not in keys_:
         raise InvalidConfig("Missing version")
-    # Accepts versions from 0.2 to 0.3 inclusive
-    elif not LooseVersion('0.2') <= ver < LooseVersion('0.4'):
+    # Accepts versions from 0.2 to 0.4 inclusive
+    elif not LooseVersion('0.2') <= ver < LooseVersion('0.5'):
         raise InvalidConfig("Unknown version")
     elif not keys_.issubset(set(['version', 'runs',
                                  'packages', 'other_files',
@@ -108,6 +108,13 @@ def load_config(filename, canonical, File=File, Package=Package):
     runs = config.get('runs', [])
     packages = read_packages(config.get('packages', []), File, Package)
     other_files = read_files(config.get('other_files', []), File)
+
+    # Adds 'input_files' and 'output_files' keys to runs
+    for run in runs:
+        if 'input_files' not in run:
+            run['input_files'] = {}
+        if 'output_files' not in run:
+            run['output_files'] = {}
 
     if canonical:
         if 'additional_patterns' in config: