+++ reprounzip/reprounzip/unpackers/default.py
@@ -25,10 +25,11 @@ import subprocess
 import sys
 import tarfile
 
+from reprounzip.common import load_config as load_config_file
 from reprounzip.unpackers.common import THIS_DISTRIBUTION, load_config, \
     select_installer, target_must_exist, shell_escape, busybox_url, \
     join_root, PKG_NOT_INSTALLED, COMPAT_OK, COMPAT_NO
-from reprounzip.utils import unicode_, download_file
+from reprounzip.utils import unicode_, iteritems, itervalues, download_file
 
 
 def installpkgs(args):
@@ -97,6 +98,9 @@ def directory_create(args):
 
     Only the files that are not part of a package are copied (unless they are
     missing from the system and were packed).
+
+    In addition, input files are put in a tar.gz (so they can be put back after
+    an upload) and the configuration file is extracted.
     """
     if not args.pack:
         sys.stderr.write("Error: setup needs --pack\n")
@@ -135,6 +139,11 @@ def directory_create(args):
         if linkname.is_absolute:
             m.linkname = join_root(root, PosixPath(m.linkname)).path
     tar.extractall(str(root), members)
+
+    # Unpacks configuration file
+    member = tar.getmember('METADATA/config.yml')
+    member.name = 'config.yml'
+    tar.extract(member, str(target))
     tar.close()
 
     # Gets library paths
@@ -161,24 +170,39 @@ def directory_create(args):
                     unicode_(join_root(root,
                                        Path(run['workingdir']))))
             cmd += ' '.join('%s=%s' % (k, shell_escape(v))
-                            for k, v in run['environ'].items())
+                            for k, v in iteritems(run['environ'])
+                            if k != 'PATH')
             cmd += ' '
+
+            # PATH
+            # Get the original PATH components
             path = [PosixPath(d)
                     for d in run['environ'].get('PATH', '').split(':')]
-            path = ':'.join(unicode_(join_root(root, d)) if d.root == '/'
-                            else unicode_(d)
-                            for d in path)
+            # The same paths but in the directory
+            dir_path = [join_root(root, d)
+                        for d in path
+                        if d.root == '/']
+            # Rebuild string
+            path = ':'.join(unicode_(d) for d in (dir_path + path))
             cmd += 'PATH=%s ' % shell_escape(path)
+
             # FIXME : Use exec -a or something if binary != argv[0]
             cmd += ' '.join(
                     shell_escape(a)
                     for a in run['argv'])
             fp.write('%s\n' % cmd)
+    (target / 'script.sh').chmod(0o755)
+
+    # Original input files, so upload can restore them
+    if any(run['input_files'] for run in runs):
+        inputtar = tarfile.open(str(target / 'inputs.tar.gz'), 'w:gz')
+        for run in runs:
+            for ifile in itervalues(run['input_files']):
+                inputtar.add(str(PosixPath(ifile)))
+        inputtar.close()
 
     # Meta-data for reprounzip
     write_dict(target / '.reprounzip', {}, 'directory')
-
-    print("Experiment set up, run %s to start" % (target / 'script.sh'))
 
 
 @target_must_exist
@@ -191,11 +215,41 @@ def directory_destroy(args):
     target.rmtree()
 
 
+def should_restore_owner(param):
+    if os.getuid() != 0:
+        if param is True:
+            # Restoring the owner was explicitely requested
+            sys.stderr.write("Error: Not running as root, cannot restore "
+                             "files' owner/group\n")
+            sys.exit(1)
+        elif param is None:
+            # Nothing was requested
+            sys.stderr.write("Warning: Not running as root, won't restore "
+                             "files' owner/group\n")
+        else:
+            # If False: skip warning
+            return False
+    else:
+        if param is None:
+            # Nothing was requested
+            sys.stderr.write("Info: Running as root, we will restore files' "
+                             "owner/group\n")
+            return True
+        elif param is True:
+            return True
+        else:
+            # If False: skip warning
+            return False
+
+
 def chroot_create(args):
     """Unpacks the experiment in a folder so it can be run with chroot.
 
     All the files in the pack are unpacked; system files are copied only if
     they were not packed, and busybox is installed if /bin/sh wasn't packed.
+
+    In addition, input files are put in a tar.gz (so they can be put back after
+    an upload) and the configuration file is extracted.
     """
     if not args.pack:
         sys.stderr.write("Error: setup/create needs --pack\n")
@@ -212,26 +266,7 @@ def chroot_create(args):
         sys.exit(1)
 
     # We can only restore owner/group of files if running as root
-    restore_owner = False
-    if os.getuid() != 0:
-        if args.restore_owner is True:
-            # Restoring the owner was explicitely requested
-            sys.stderr.write("Error: Not running as root, cannot restore "
-                             "files' owner/group\n")
-            sys.exit(1)
-        elif args.restore_owner is None:
-            # Nothing was requested
-            sys.stderr.write("Warning: Not running as root, won't restore "
-                             "files' owner/group\n")
-        # If False: skip warning
-    else:
-        if args.restore_owner is None:
-            # Nothing was requested
-            sys.stderr.write("Info: Running as root, we will restore files' "
-                             "owner/group\n")
-            restore_owner = True
-        elif args.restore_owner is True:
-            restore_owner = True
+    restore_owner = should_restore_owner(args.restore_owner)
 
     # Loads config
     runs, packages, other_files = load_config(pack)
@@ -282,6 +317,11 @@ def chroot_create(args):
             m.uid = uid
             m.gid = gid
     tar.extractall(str(root), members)
+
+    # Unpacks configuration file
+    member = tar.getmember('METADATA/config.yml')
+    member.name = 'config.yml'
+    tar.extract(member, str(target))
     tar.close()
 
     # Sets up /bin/sh and /usr/bin/env, downloading busybox if necessary
@@ -307,7 +347,7 @@ def chroot_create(args):
             cmd = 'cd %s && ' % shell_escape(run['workingdir'])
             cmd += '/usr/bin/env -i '
             cmd += ' '.join('%s=%s' % (k, shell_escape(v))
-                            for k, v in run['environ'].items())
+                            for k, v in iteritems(run['environ']))
             cmd += ' '
             # FIXME : Use exec -a or something if binary != argv[0]
             cmd += ' '.join(
@@ -318,11 +358,18 @@ def chroot_create(args):
                      userspec,
                      shell_escape(unicode_(root)),
                      shell_escape(cmd)))
+    (target / 'script.sh').chmod(0o755)
+
+    # Original input files, so upload can restore them
+    if any(run['input_files'] for run in runs):
+        inputtar = tarfile.open(str(target / 'inputs.tar.gz'), 'w:gz')
+        for run in runs:
+            for ifile in itervalues(run['input_files']):
+                inputtar.add(str(PosixPath(ifile)))
+        inputtar.close()
 
     # Meta-data for reprounzip
     write_dict(target / '.reprounzip', {}, 'chroot')
-
-    print("Experiment set up, run %s to start" % (target / 'script.sh'))
 
 
 @target_must_exist
@@ -364,6 +411,155 @@ def run(args):
     read_dict(target / '.reprounzip', args.type)
 
     subprocess.check_call(['/bin/sh', (target / 'script.sh').path])
+
+
+@target_must_exist
+def upload(args):
+    """Replaces an input file in the directory.
+    """
+    target = Path(args.target[0])
+    files = args.file
+    unpacked_info = read_dict(target / '.reprounzip', args.type)
+    input_files = unpacked_info.setdefault('input_files', {})
+
+    # Loads config
+    runs, packages, other_files = load_config_file(target / 'config.yml',
+                                                   True)
+
+    # No argument: list all the input files and exit
+    if not files:
+        print("Input files:")
+        for i, run in enumerate(runs):
+            if len(runs) > 1:
+                print("  Run %d:" % i)
+            for input_name in run['input_files']:
+                assigned = input_files.get(input_name) or "(original)"
+                print("    %s: %s" % (input_name, assigned))
+        return
+
+    restore_owner = False
+    if args.type == 'chroot':
+        restore_owner = should_restore_owner(args.restore_owner)
+
+    root = (target / 'root').absolute()
+
+    # Get the path of each input file
+    all_input_files = {}
+    for run in runs:
+        all_input_files.update(run['input_files'])
+
+    try:
+        # Copy files
+        for filespec in files:
+            filespec_split = filespec.rsplit(':', 1)
+            if len(filespec_split) != 2:
+                sys.stderr.write("Invalid file specification: %r\n" % filespec)
+                sys.exit(1)
+            local_path, input_name = filespec_split
+
+            try:
+                input_path = PosixPath(all_input_files[input_name])
+            except KeyError:
+                sys.stderr.write("Invalid input name: %r" % input_name)
+                sys.exit(1)
+
+            remote_path = join_root(root, input_path)
+
+            temp = None
+
+            if not local_path:
+                # Restore original file from pack
+                fd, temp = Path.tempfile(prefix='reprozip_input_')
+                os.close(fd)
+                tar = tarfile.open(str(target / 'inputs.tar.gz'), 'r:*')
+                member = tar.getmember(str(
+                        join_root(PosixPath(''), input_path)))
+                member.name = str(temp.name)
+                tar.extract(member, str(temp.parent))
+                tar.close()
+                local_path = temp
+            else:
+                local_path = Path(local_path)
+                if not local_path.exists():
+                    sys.stderr.write("Local file %s doesn't exist\n" %
+                                     local_path)
+                    sys.exit(1)
+
+            # Copy
+            orig_stat = remote_path.stat()
+            local_path.copyfile(remote_path)
+            remote_path.chmod(orig_stat.st_mode & 0o7777)
+            if restore_owner:
+                remote_path.chown(orig_stat.st_uid, orig_stat.st_gid)
+
+            if temp is not None:
+                temp.remove()
+                input_files[input_name] = None
+            else:
+                input_files[input_name] = local_path.absolute().path
+    finally:
+        write_dict(target / '.reprounzip', unpacked_info, args.type)
+
+
+@target_must_exist
+def download(args):
+    """Gets an output file from the directory.
+    """
+    target = Path(args.target[0])
+    files = args.file
+    read_dict(target / '.reprounzip', args.type)
+
+    # Loads config
+    runs, packages, other_files = load_config_file(target / 'config.yml',
+                                                   True)
+
+    # No argument: list all the output files and exit
+    if not files:
+        print("Output files:")
+        for i, run in enumerate(runs):
+            if len(runs) > 1:
+                print("  Run %d:" % i)
+            for output_name in run['output_files']:
+                print("    %s" % output_name)
+        return
+
+    root = (target / 'root').absolute()
+
+    # Get the path of each output file
+    all_output_files = {}
+    for run in runs:
+        all_output_files.update(run['output_files'])
+
+    # Copy files
+    for filespec in files:
+        filespec_split = filespec.split(':', 1)
+        if len(filespec_split) != 2:
+            sys.stderr.write("Invalid file specification: %r\n" % filespec)
+            sys.exit(1)
+        output_name, local_path = filespec_split
+
+        try:
+            remote_path = PosixPath(all_output_files[output_name])
+        except KeyError:
+            sys.stderr.write("Invalid output name: %r\n" % output_name)
+            sys.exit(1)
+
+        remote_path = join_root(root, remote_path)
+
+        if not local_path:
+            # Output to stdout
+            with remote_path.open('rb') as fp:
+                chunk = fp.read(1024)
+                if chunk:
+                    sys.stdout.write(chunk)
+                while len(chunk) == 1024:
+                    chunk = fp.read(1024)
+                    if chunk:
+                        sys.stdout.write(chunk)
+        else:
+            # Copy
+            remote_path.copyfile(local_path)
+            remote_path.copymode(local_path)
 
 
 def test_same_pkgmngr(pack, config, **kwargs):
@@ -437,14 +633,23 @@ def setup_directory(parser):
     parser_setup.add_argument('--pack', nargs=1, help="Pack to extract")
     parser_setup.set_defaults(func=directory_create)
 
-    # TODO : directory upload
+    # upload
+    parser_upload = subparsers.add_parser('upload', parents=[options])
+    parser_upload.add_argument('file', nargs=argparse.ZERO_OR_MORE,
+                               help="<path>:<input_file_name>")
+    parser_upload.set_defaults(func=upload, type='directory')
 
     # run
     parser_run = subparsers.add_parser('run', parents=[options])
     parser_run.add_argument('run', default=None, nargs='?')
     parser_run.set_defaults(func=run, type='directory')
 
-    # TODO : directory download
+    # download
+    parser_download = subparsers.add_parser('download',
+                                            parents=[options])
+    parser_download.add_argument('file', nargs=argparse.ZERO_OR_MORE,
+                                 help="<output_file_name>:<path>")
+    parser_download.set_defaults(func=download, type='directory')
 
     # destroy
     parser_destroy = subparsers.add_parser('destroy', parents=[options])
@@ -513,14 +718,24 @@ def setup_chroot(parser):
             help=argparse.SUPPRESS)
     parser_setup.set_defaults(func=chroot_setup)
 
-    # TODO : chroot upload (options, opt_owner)
+    # upload
+    parser_upload = subparsers.add_parser('upload',
+                                          parents=[options, opt_owner])
+    parser_upload.add_argument('file', nargs=argparse.ZERO_OR_MORE,
+                               help="<path>:<input_file_name>")
+    parser_upload.set_defaults(func=upload, type='chroot')
 
     # run
     parser_run = subparsers.add_parser('run', parents=[options])
     parser_run.add_argument('run', default=None, nargs='?')
     parser_run.set_defaults(func=run, type='chroot')
 
-    # TODO : chroot download (options)
+    # download
+    parser_download = subparsers.add_parser('download',
+                                            parents=[options])
+    parser_download.add_argument('file', nargs=argparse.ZERO_OR_MORE,
+                                 help="<output_file_name>:<path>")
+    parser_download.set_defaults(func=download, type='chroot')
 
     # destroy
     parser_destroy = subparsers.add_parser('destroy', parents=[options])